Drop CustodyProvider.issuer; rename account-mapping file to account-resolver
The issuer is the env-injected signer (ASSET_ISSUER_PRIVATE_KEY), never a
custody wallet, so the custody interface loses its issuer slot:

- CustodyProvider.issuer removed; rpcProvider now derives from the escrow
  wallet in both providers.
- DirectTokenService takes the env issuer wallet via constructor (built at
  the composition root; ephemeral fallback with a warning until the
  plugins accept an optional issuer) and uses it for deploy, mint and
  balanceOf.
- Fireblocks/DFNS providers and configs stop reading
  ASSET_ISSUER_CUSTODY_ACCOUNT_ID and constructing issuer wallets; the
  Fireblocks base vault and the DFNS config signer derive from escrow.

Also rename accounts/account-mapping.ts -> account-resolver.ts (with the
test and jest config) to align with the AccountResolver naming.
