+++ package.json
@@ -17,10 +17,10 @@
     "clean": "rm -rf dist",
     "prebuild": "eslint ./src --ext .ts --fix",
     "build": "tsc",
-    "test": "jest && jest --config jest.account-mapping.config.js && jest --config jest.omnibus.config.js && jest --config jest.legacy-migration.config.js && jest --config jest.plan-approval.config.js && jest --config jest.operation-context.config.js",
+    "test": "jest && jest --config jest.account-resolver.config.js && jest --config jest.omnibus.config.js && jest --config jest.legacy-migration.config.js && jest --config jest.plan-approval.config.js && jest --config jest.operation-context.config.js",
     "test:fireblocks": "jest --config jest.fireblocks.config.js",
     "test:dfns": "jest --config jest.dfns.config.js",
-    "test:account-mapping": "jest --config jest.account-mapping.config.js",
+    "test:account-mapping": "jest --config jest.account-resolver.config.js",
     "test:omnibus": "jest --config jest.omnibus.config.js",
     "test:legacy-migration": "jest --config jest.legacy-migration.config.js",
     "test:deposits-omnibus": "jest --config jest.deposits-omnibus.config.js",
