+++ src/services/accounts/index.ts
@@ -1,3 +1,3 @@
-export * from './account-mapping'
+export * from './account-resolver'
 export * from './mapping-validator'
 export * from './mapping-config'
