+++ src/services/plan-approval/options/build-custody-plan-approval.ts
@@ -3,7 +3,7 @@ import { FinP2PClient } from "@owneraio/finp2p-client";
 import { ConfigurablePlanApprovalService } from "../configurable-plan-approval-service";
 import { PlanApprovalOption } from "../option";
 import { CustodyProvider } from "../../custody/custody-provider";
-import { AccountResolver, AssetStore } from "../../accounts/account-mapping";
+import { AccountResolver, AssetStore } from "../../accounts/account-resolver";
 import { isHederaNetwork } from "../../custody/wallet-activation";
 import { TokenWhitelistingOption } from "./token-whitelisting-option";
 import { GasPrefundingOption } from "./gas-prefunding-option";
